리샘플러 초기화 시 데이터 타입 적용: MambaCompressor 초기화 시 dtype 매개변수 추가

+++ src/models/build.py
@@ -67,7 +67,7 @@ class CustomVLMModel(PreTrainedModel):
         # 4) Optional resampler ---------------------------------------------
         if getattr(config, "use_resampler", False):
             from resampler.mamba_ssm.modules.mamba_compressor import MambaCompressor
-            self.resampler = MambaCompressor(d_model=d_v, n_layer=1, fp32=False)
+            self.resampler = MambaCompressor(d_model=d_v, n_layer=1, fp32=False).to(dtype)
         else:
             self.resampler = None
